refactor: Rename Configurator to Configure
