+++ src/cleantest/control/__init__.py
@@ -2,4 +2,4 @@
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-from .configurator import Configure
+from .configure import Configure
